requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
spacy==3.7.2
loguru==0.7.2
pydantic-settings==2.2.1
